// oldSrc/types/contentful.ts
import type { IGatsbyImageData } from 'gatsby-plugin-image'
import type { ContentfulRichTextGatsbyReference, RenderRichTextData } from 'gatsby-source-contentful/rich-text'
import type { JsonLdType } from './jsonld'

interface ExcerptListReference extends ContentfulRichTextGatsbyReference {
    limit: number
}

interface RichTextPostReference extends ContentfulRichTextGatsbyReference {
    slug: string
}

interface ImageWithCaptionReference extends ContentfulRichTextGatsbyReference {
    caption: string
    image: {
        gatsbyImageData: IGatsbyImageData
        description: string
    }
}

interface ImageReference extends ContentfulRichTextGatsbyReference {
    description: string
    gatsbyImageData: IGatsbyImageData
}

export interface ContactInfoLink {
    contentful_id: string
    title: string
    url: string
    icon: 'facebook' | 'bluesky' | 'instagram' | 'linkedin' | 'envelope' | 'threads' | 'mastodon' | 'tiktok'
}

export interface Fiduciary {
    duty: string
    organization: string
    startYear: number
    endYear?: number
}

export interface JobExperience {
    title: string
    company: string
    location: string
    startYear: number
    endYear?: number
}

export interface Degree {
    degree: string
    school: string
    location: string
    startYear: number
    endYear?: number
}

interface CurriculumVitae extends ContentfulRichTextGatsbyReference {
    fiduciariesTitle: string
    fiduciaries: Fiduciary[]
    degreesTitle: string
    degrees: Degree[]
    jobExperiencesTitle: string
    jobExperiences: JobExperience[]
}

interface ContactInfoLinkReference extends ContentfulRichTextGatsbyReference {
    title: string
    url?: string | null
    icon: 'facebook' | 'bluesky' | 'instagram' | 'linkedin' | 'envelope' | 'threads' | 'mastodon' | 'tiktok'
}

type RichReferences =
    | ExcerptListReference
    | RichTextPostReference
    | ImageWithCaptionReference
    | ContactInfoLinkReference
    | ContentfulRichTextGatsbyReference
    | CurriculumVitae
    | ImageReference

export type RichBody = RenderRichTextData<RichReferences>

export interface ContentfulPage {
    description: string
    jsonLdType: JsonLdType
    title: string
    subtitle: string | null
    secondaryTitle: string | null
    body: RichBody
    image: {
        gatsbyImageData: IGatsbyImageData
        description: string
    }
    socialImage: {
        gatsbyImageData: IGatsbyImageData
    } | null
    backgroundImage: {
        gatsbyImageData: IGatsbyImageData
        description: string
    } | null
    updatedAt: string
    slug: string
}

interface ContentfulPostCommon {
    title: string
    createdAt: string
    publishDate?: string | null
    metadata: {
        tags: {
            contentful_id: string
            name: string
        }[]
    }
    headerImage: {
        gatsbyImageData: IGatsbyImageData
        description: string
    }
    backgroundImage?: {
        gatsbyImageData: IGatsbyImageData
        description: string
    } | null
    socialImage?: {
        gatsbyImageData: IGatsbyImageData
    } | null
    excerpt: string
}

export interface ContentfulPostExcerpt extends ContentfulPostCommon {
    slug: string
}

export interface ContentfulPost extends ContentfulPostCommon {
    body: RichBody
    updatedAt: string
    publishedOld: string | null
    published: string
}

export interface MainNav {
    links: {
        contentful_id: string
        title: string
        navigationTitle?: string
        slug: string
    }[]
}

export interface FooterNav {
    links: ContactInfoLink[]
}

export interface Tag {
    contentful_id: string
    name: string
}